requirements.txt
## Server Dependencies
uvicorn[standard]
gunicorn

## Web Frameworks
fastapi
pydantic

## Project Dependencies
Pillow
requests
aiofiles
